requirements.txt
mkdocs
mkdocs-cinder
cookiecutter
pytest
black
attrs
